+++ example/block_aware/block_aware_app.go
@@ -55,9 +55,8 @@ func (app *ChainAwareApplication) Query(reqQuery types.RequestQuery) (resQuery t
 	}
 }
 
-func (app *ChainAwareApplication) BeginBlock(hash []byte, header *types.Header) {
+func (app *ChainAwareApplication) BeginBlock(reqBeginBlock types.RequestBeginBlock) {
 	app.beginCount++
-	return
 }
 
 func (app *ChainAwareApplication) EndBlock(height uint64) (resEndBlock types.ResponseEndBlock) {