+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.6.0 (September 22, 2017)
+
+BREAKING CHANGES:
+
+- [types/client] app.BeginBlock takes RequestBeginBlock
+- [types/client] app.InitChain takes RequestInitChain
+- [types/client] app.Info takes RequestInfo
+
+IMPROVEMENTS:
+
+- various linting
+
 ## 0.5.0 (May 18, 2017)
 
 BREAKING CHANGES: